move head hash method
